added fileExists function and one-time setup conf file

+++ Jenkinsfile
@@ -13,33 +13,39 @@ pipeline {
                         }
                     }
                 }
-                dir("${JENKINS_HOME}/workspace") {
+                dir("${JENKINS_HOME}/workspace/n8n-setup") {
                     script {
-                        if ( !fileExists( '/etc/apt/apt.conf.d/00-proxy' ) ) {
-                            sh 'echo "Acquire::http::proxy \\"http://proxy-dmz.intel.com:911\\";\nAcquire::https::proxy \\"http://proxy-dmz.intel.com:912\\";" >> /etc/apt/apt.conf.d/00-proxy'           
-                        }
-                        sh 'apt-get update -y'
-                        node_res = sh( script: 'test -e ./node-v14.18.0-linux-x64.tar.gz', returnStatus: true ) == 0
-                        if ( node_res == false ) {
-                            sh 'curl -O https://nodejs.org/dist/v14.18.0/node-v14.18.0-linux-x64.tar.gz'
-                        }
-                        sh "mkdir -p /usr/local/lib/nodejs"
-                        dir_empty = sh( script: 'test -z "$(ls -A /usr/local/lib/nodejs)"', returnStatus: true ) == 0
-                        if ( dir_empty == true ) {
-                            sh 'apt-get install -y tar'
-                            sh 'apt-get install -y gzip'
-                            sh "tar -xvzf node-v14.18.0-linux-x64.tar.gz -C /usr/local/lib/nodejs"
-                        }
-                        sh 'apt-get install -y build-essential python'
-                        npm_res = sh( script: 'test -f /usr/bin/npm', returnStatus: true ) == 0
-                        if ( npm_res == false ) {
-                            sh "ln -s /usr/local/lib/nodejs/nodejs/node-v14.18.0-linux-x64/bin/npm /usr/bin/npm"
-                            sh "ln -s /usr/local/lib/nodejs/node-v14.18.0-linux-x64/bin/node /usr/bin/node"
-                        }
-                        lerna_res = sh( script: 'test -f /usr/bin/lerna', returnStatus: true ) == 0
-                        if ( result == false ) {
-                            sh 'npm install -g lerna'
-                            sh "ln -s /usr/local/lib/nodejs/node-v14.18.0-linux-x64/bin/lerna /usr/bin/lerna"
+                        if ( !fileExist ('./setup.conf')) {
+
+                            if ( !fileExists( '/etc/apt/apt.conf.d/00-proxy' ) ) {
+                                sh 'echo "Acquire::http::proxy \\"http://proxy-dmz.intel.com:911\\";\nAcquire::https::proxy \\"http://proxy-dmz.intel.com:912\\";" >> /etc/apt/apt.conf.d/00-proxy'           
+                            }
+                            
+                            sh 'apt-get update -y'
+                            
+                            if ( !fileExists ('./node-v14.18.0-linux-x64.tar.gz') ) {
+                                sh 'curl -O https://nodejs.org/dist/v14.18.0/node-v14.18.0-linux-x64.tar.gz'
+                            }
+                            
+                            sh "mkdir -p /usr/local/lib/nodejs"
+                            dir_empty = sh( script: 'test -z "$(ls -A /usr/local/lib/nodejs)"', returnStatus: true ) == 0
+                            if ( dir_empty == true ) {
+                                sh 'apt-get install -y tar'
+                                sh 'apt-get install -y gzip'
+                                sh "tar -xvzf node-v14.18.0-linux-x64.tar.gz -C /usr/local/lib/nodejs"
+                            }
+
+                            sh 'apt-get install -y build-essential python'
+
+                            if ( !fileExists ('/usr/bin/npm') ) {
+                                sh "ln -s /usr/local/lib/nodejs/nodejs/node-v14.18.0-linux-x64/bin/npm /usr/bin/npm"
+                                sh "ln -s /usr/local/lib/nodejs/node-v14.18.0-linux-x64/bin/node /usr/bin/node"
+                            }
+                            
+                            if ( !fileExists ('/usr/bin/lerna') ) {
+                                sh 'npm install -g lerna'
+                                sh "ln -s /usr/local/lib/nodejs/node-v14.18.0-linux-x64/bin/lerna /usr/bin/lerna"
+                            }
                         }
                     }
                 }
@@ -48,8 +54,8 @@ pipeline {
                         UPDATED = false
                         if ( CLONE == false ) {
                             sh 'git remote update'
-                            LOCAL = sh( script: 'git rev-parse @', returnStdout: true )
-                            REMOTE = sh( script: 'git rev-parse @{u}', returnStdout: true )
+                            LOCAL = sh( script: 'git rev-parse @', returnStdout: true ).trim()
+                            REMOTE = sh( script: 'git rev-parse @{u}', returnStdout: true ).trim()
                             if ( LOCAL != REMOTE ) {
                                 sh 'git pull origin master'
                                 UPDATED = true
@@ -59,6 +65,10 @@ pipeline {
                             sh 'lerna bootstrap --hoist'
                             sh 'npm run build'
                             sh './packages/cli/bin/n8n --version'
+
+                            if ( !fileExist ("${JENKINS_HOME}/workspace/n8n-setup/setup.conf")) {
+                                sh 'echo "DONE" >> ${JENKINS_HOME}/workspace/n8n-setup/setup.conf'
+                            }
                         }
                     }
                 }
